[Feat] KeynoteViewAdapter 정의!!
